% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/direction_summary.R
\name{direction_summary}
\alias{direction_summary}
\title{Summary of movement direction for Oregon RFID antenna data}
\usage{
direction_summary(dir_df, include_stationary = FALSE)
}
\arguments{
\item{dir_df}{output from \code{\link{tag_direction}}.}

\item{include_stationary}{if \emph{TRUE}, all detections are summarized. If FALSE, only detections with a known movement direction (up or down) are included in the summary.}
}
\value{
Returns a tibble object.
}
\description{
Summarizes the time difference between the first and last movement directions for each unique tag ID. Input data are created by \code{\link{tag_direction}}, which determines the direction of movement for each detection event in \code{\link{join_multireader_data}}.
}
\details{
\code{direction_summary} isolates the first and last direction of movement from \code{\link{tag_direction}} and determines the time difference in seconds and days. Directions are U for upstream movement, D for downstream movement, and S for no movement, or consecutive detection at the same location.

It is common for a tag to be detected multiple times at the same antenna, which will result in a movement direction of S, or stationary. The user is encouraged to examine direction summaries for \emph{include_stationary = TRUE} and \emph{include_stationary = FALSE} to become familiar with their data. When \emph{include_stationary = FALSE}, there may be fewer tag ids in the direction summary than in the full data set.
}
\examples{

# Create a list containing compiled reader data:
readers <- list(reader_us, reader_ds)

# Join data into a multi-reader array:
PIT_data <- join_multireader_data(readers)

# List readers:
unique(PIT_data$LOC)

# Determine tag direction for animals moving from downstream to upstream:
dir <- tag_direction(PIT_data, c("downstream_A1", "upstream_A1"))
 
# Determine the time difference between first and last detections:
dir_summary <- direction_summary(dir)

}
\seealso{
\code{\link{import_ORFID}} for importing data files from Oregon RFID ORMR and ORSR antenna readers.

\code{\link{join_multireader_data}} for combining data from Oregon RFID ORMR and ORSR antenna readers into a multi-reader array.

\code{\link{tag_direction}} for determining movement direction of detections in systems with a linear migration route.
}
\author{
Annika Putt <annika@instream.net>
}
