% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_ORFID.R
\name{import_ORFID}
\alias{import_ORFID}
\title{Import data files from the new generation of Oregon RFID readers}
\usage{
import_ORFID(file, delim)
}
\arguments{
\item{file}{txt file generated by an Oregor RFID reader.}

\item{delim}{the field separator character. Values on each line of the file are separated by this character and it must be either tab, comma or semi-colon.}
}
\value{
Returns a tibble object.
}
\description{
Function to import files recorded by the new generation of Oregon RFID readers.
}
\details{
A data frame is created in the user environment from a import of a Oregor RFID reader file.
  
the field separator character must be either tab, comma or semi-colon otherwise the function cannot organize properly the variables, generating error messages, without import the file.

The Tag number (TAG) is required for further estatistical analysis. Consider in upload the data from the ORFID reader with variable selected in the Reader field names.
}
\examples{

##  Importing the ORFID reader data to the user working directory

\dontrun{} data_BRA <- import_ORFID("~/Desktop/SCD_BRA.txt")

\dontrun{} data_USA <- import_ORFID("~/Desktop/SCD_USA.txt")
}
\author{
Hugo Marques <biohmarques@gmail.com>
}
