% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_ORFID_data.R
\name{join_ORFID_data}
\alias{join_ORFID_data}
\title{Combine data from ORFID readers into an multi readers array}
\usage{
join_ORFID_data(...)
}
\arguments{
\item{...}{A list or a set of dataframes containing data frames created with the \code{\link{import_ORFID}} function to be combined.}
}
\value{
Returns a tibble object.
}
\description{
Function allows users to combine unique readers into an array, using the data obtained from \code{\link{import_ORFID}} function. It is necessary to create a list with all data frames to be joined. If the data frames have diferent variables,
}
\details{
Data frame is created in the user environment. The output of \code{\link{join_ORFID_data}} will contain a column if that column appears in any of data frames combined. A factor class variable called LOC (from location) is the combination of SCD (Site Code) and ANT (Antenna) variables. It should be used as location in further statistical analysis once it is the individual detection spot.
}
\examples{
\dontrun{

##  Create the list containing the imported files:
readers <- list(reader_1, reader_2, reader_3)

##  Combine the files:
PIT_data <- join_ORFID_data(readers)

}
}
\seealso{
\code{\link{import_ORFID}} for importing data files from the new generation of Oregon RFID readers
}
\author{
Hugo Marques <biohmarques@gmail.com>
}
