% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marker_tag.R
\name{marker_tag}
\alias{marker_tag}
\title{Summarize marker tag data from \code{\link{import_ORFID}}}
\usage{
marker_tag(x, tag, gap)
}
\arguments{
\item{x}{antenna data compiled using \code{\link{import_ORFID}} or \code{\link{join_multireader_data}}.}

\item{tag}{marker tag identification (character object).}

\item{gap}{minimum time gap in seconds (optional)}
}
\value{
Returns a tibble object.
}
\description{
Summarize marker tag data from \code{\link{import_ORFID}}
}
\details{
A data frame is created containing only information from the specified marker tag.
}
\examples{
\dontrun{

# Create a list containing compiled reader data:
readers <- list(reader_1, reader_2, reader_3)

# Join data into a multi-reader array:
PIT_data <- join_multireader_data(readers)

# Summarize marker tag data
marker_tag(PIT_data)
}
}
\seealso{
\code{\link{import_ORFID}} for importing data files from Oregon RFID ORMR and ORSR antenna readers.

\code{\link{join_multireader_data}} for combining data from Oregon RFID ORMR and ORSR antenna readers into a multi-reader array.
}
\author{
Hugo Marques <biohmarques@gmail.com>
}
