% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marker_tag_plot.R
\name{marker_tag_plot}
\alias{marker_tag_plot}
\title{Plot detection from marker tags used with Oregon RFID ORMR and ORSR antenna readers}
\usage{
marker_tag_plot(x, tag, gap)
}
\arguments{
\item{x}{antenna data compiled using \code{\link{import_ORFID}} or \code{\link{join_multireader_data}}.}

\item{tag}{marker tag identification (character object).}

\item{gap}{minimum time gap in seconds (optional)}
}
\value{
Returns a plot object.
}
\description{
Creates a time series plot containing only detection data from a user-specified marker tag (unique tag number).
}
\details{
Creates a plot displaying marker tag detections.
}
\examples{
\dontrun{

# Create a list containing compiled reader data:
readers <- list(reader_1, reader_2, reader_3)

# Join data into a multi-reader array:
PIT_data <- join_multireader_data(readers)

# Plot marker tag data:
marker_tag_plot(PIT_data, "0000_000000004978", gap = 360)
}
}
\seealso{
\code{\link{import_ORFID}} for importing data files from Oregon RFID ORMR and ORSR antenna readers.

\code{\link{join_multireader_data}} for combining data from Oregon RFID ORMR and ORSR antenna readers into a multi-reader array.
}
\author{
Hugo Marques <biohmarques@gmail.com>
}
