% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ORFID_import.R
\name{ORFID_import}
\alias{ORFID_import}
\title{Import data files from the new generation of Oregon RFID readers}
\usage{
ORFID_import(file, delim = "\\t", skip = NULL)
}
\arguments{
\item{file}{txt file generated by an Oregor RFID reader.}

\item{delim}{the field separator character. Values on each line of the file are separated by this character.}

\item{skip}{integer: the number of lines of the data file to skip before beginning to read data.}
}
\description{
Function to import files recorded by the new generation of Oregon RFID readers.
}
\details{
Data frame is saved in the user-specified working directory.
}
\examples{
#save data frame to a working directory
\dontrun{} ORFID_import("~/Desktop/SCD_BRA.txt")
}
