% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/field_names.R
\name{ORFID_names}
\alias{ORFID_names}
\title{Returns the field names from ORSR and ORMR readers}
\usage{
ORFID_names()
}
\value{
A data frame with two variables is returned (Name and Details)
}
\description{

}
\details{
Returns the field names and its details from ORSR and ORMR readers data
}
\examples{

## Creating a data frame containing the field names from ORSR and ORMR readers.

(FM <- field_names())
 
}
\author{
Hugo Marques <biohmarques@gmail.com>
}
