% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{reader_3}
\alias{reader_3}
\title{ORFID readers data samples}
\format{A data frame with xxx rows and xx variables:
\describe{
  \item{DTY}{Detection type (S = summary, I = individual, E = event)}
  \item{ARR}{Arrival date and time}
  \item{DUR}{Duration of the period between the arrival and the departure from the detection zone}
  \item{TTY}{Tag type A=Animal (R = Read only, W = Writeable, P = Phantom)}
  \item{TAG}{Tag ID number}
  \item{SCD}{Site code}
  \item{NCD}{Number of consecutive detections}
}}
\usage{
reader_3
}
\description{
Dataset to illustrate the ORFID functions. The variables are as follows:
}
\keyword{datasets}
