% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{reader_us}
\alias{reader_us}
\title{ORFID data samples}
\format{
A data frame with 781 rows and 16 variables:
\describe{
  \item{DTY}{Detection type (S = summary, I = individual, E = event)}
  \item{ARR}{Arrival date and time}
  \item{TRF}{Time reference (G = GNSS, N = network, U = unreferenced)}
  \item{DUR}{Duration of the period between the arrival and the departure from the detection zone}
  \item{TTY}{Tag type A=Animal (R = Read only, W = Writeable, P = Phantom)}
  \item{TAG}{Tag ID number}
  \item{SCD}{Site code}
  \item{NCD}{Number of consecutive detections}
  \item{EFA}{Effective amps}
  \item{TCH}{Tag technology (HDX, FDX, HF)}
  \item{ANT}{Antenna number}
  \item{EMP}{Number of empty scans preceding detection}
  \item{TSS}{Tag signal strength}
  \item{SPV}{Unknown}
  \item{NOI}{Noise}
  \item{CLS}{Tag class}
}
}
\usage{
reader_us
}
\description{
Compiled data to illustrate directional functions from \pkg{ORFID}. The variables are as follows:
}
\keyword{datasets}
