% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TAG_info.R
\name{TAG_info}
\alias{TAG_info}
\title{Summarizing tag-specific info of \code{\link{import_ORFID}}.}
\usage{
TAG_info(x)
}
\arguments{
\item{x}{PIT_data data frame to analyzed.}
}
\value{
Returns a tibble object.
}
\description{
Function for getting info about the tags recorded in a PIT_data data frame from \code{\link{import_ORFID}} function (number of unique tags, tags type and records per tag).
}
\details{
DAta frame is created in the user environment.
}
\examples{

#Analyzing tag-specific information in a PIT_data.#' 
\dontrun{} tags <- TAG_info(PIT_data)

# How many tags detected:#' 
\dontrun{} lenght(tags$TAG)

}
\author{
Hugo Marques
}
