% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tags_summary.R
\name{tags_summary}
\alias{tags_summary}
\title{Summarizing tag-specific info of \code{\link{import_ORFID}}}
\usage{
tags_summary(x)
}
\arguments{
\item{x}{data frame to be analyzed.}
}
\value{
Returns a tibble object.
}
\description{
Function for getting info about the tags recorded in a PIT_data data frame from \code{\link{import_ORFID}} function (number of unique tags, tags type and records per tag).
}
\details{
A data frame is created in the user environment with the data grouped by TAG IDs.
}
\examples{
##  Analyzing tag-specific information in a PIT_data.
tag_info(reader_1)

##  Create the list containing the imported files:
readers list(reader_1, reader_2, reader_3)

##  Combine the files:
array <- join_multireader_data(readers)

##  Summarizing tag info
tag_info(array)
}
\seealso{
\code{\link{import_ORFID}} for importing data files from the new generation of Oregon RFID readers
\code{\link{join_multireader_data}} for Combining data from ORFID readers into an multi readers array
}
\author{
Hugo Marques <biohmarques@gmail.com>
}
