% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tags_summary1.R
\name{tags_summary1}
\alias{tags_summary1}
\title{Summarizing tag-specific info of \code{\link{import_ORFID}}}
\usage{
tags_summary1(x)
}
\arguments{
\item{x}{data frame to be analyzed.}
}
\value{
Returns a tibble object.
}
\description{
Function for getting info about the tags recorded in a PIT_data data frame from \code{\link{import_ORFID}} function (number of unique tags, tags type and records per tag).
}
\details{
A data frame is created in the user environment with the data grouped by TAG IDs.
}
\examples{
\dontrun{

##  Create the list containing the imported files:
readers <- list(reader_1, reader_2)

##  Combine the files:
PIT_data <- join_multireader_data(readers)

##  Analyzing tag-specific information in a PIT_data.
tags_summary1(PIT_data)
}
}
\seealso{
\code{\link{import_ORFID}} for importing data files from the new generation of Oregon RFID readers
\code{\link{join_multireader_data}} for Combining data from ORFID readers into an multi readers array
}
\author{
Hugo Marques <biohmarques@gmail.com>
}
