% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ant_efficiency.R
\name{ant_efficiency}
\alias{ant_efficiency}
\title{Determines detection efficiency for \code{\link{join_multireader_data}}}
\usage{
ant_efficiency(x, LOC_vec)
}
\arguments{
\item{x}{data frame to be analyzed}

\item{LOC_vec}{vector of antenna locations from first encountered to last encountered}
}
\value{
Returns a tibble object.
}
\description{
Function for determining detection efficiency for each antenna in systems where multiple antennas are used along a linear migration route.
}
\details{
\code{ant_efficiency} determines the detection efficiency of each antenna in a linear migration route. Direction is determined based on the order of locations from first encountered to last encountered, as specified in \emph{LOC_vec}. Use \code{\link{site_summary}} to identify all locations present in the multi reader data, which must be included in \emph{LOC_vec}. 

Antenna efficiency is determined by identifying which tags were detected at antenna x and which tags were detected anywhere after/above antenna x. The efficiency of antenna x is then the number of shared tag detections divided by the total number of detections after x. Note that efficiency and shared detections cannot be determined for the final antenna as there are no subsequent detections. Reversing the order of \emph{LOC_vec} can inform efficiency in systems with movement in multiple directions.
}
\examples{
\dontrun{

##  Create the list containing the imported files:
readers <- list(reader_1, reader_2)

##  Combine the files:
PIT_data <- join_multireader_data(readers)

##  Example of format for determining tag direction.
ant_efficiency(PIT_data, c("BRA", "USA"))
}
}
\seealso{
\code{\link{import_ORFID}} for importing data files from the new generation of Oregon RFID readers

\code{\link{join_multireader_data}} for combining data from ORFID readers into a multi reader array

\code{\link{site_summary}} for identifying all locations present in a multi reader array
}
\author{
Annika Putt
}
