% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/direction_summary.R
\name{direction_summary}
\alias{direction_summary}
\title{Summarizing time difference between first and last directions in \code{\link{tag_direction}}}
\usage{
direction_summary(dir_df)
}
\arguments{
\item{dir_df}{output from \code{\link{tag_direction}}}
}
\value{
Returns a tibble object.
}
\description{
Function for summarizing the time between the first and last movement directions for each unique tag id. Input data are created by \code{\link{tag_direction}}, which determines the direction of movement for each detection event in \code{\link{join_multireader_data}}.
}
\details{
\code{direction_summary} isolates the first and last direction of movement from \code{\link{tag_direction}} and determines the time difference in seconds and days.
}
\examples{
\dontrun{

# Create a list containing compiled reader data:
readers <- list(reader_1, reader_2)

# Join data into a multi-reader array:
PIT_data <- join_multireader_data(readers)

# Determine tag direction:
dir <- tag_direction(PIT_data, c("BRA", "USA"))

# Determine the time difference between first and last detections:
dir_summary <- direction_summary(dir)
}
}
\seealso{
\code{\link{import_ORFID}} for importing data files from Oregon RFID ORMR and ORSR antenna readers.

\code{\link{join_multireader_data}} for combining data from Oregon RFID ORMR and ORSR antenna readers into a multi-reader array.

\code{\link{tag_direction}} for determining movement direction of detections in systems with a linear migration route.
}
\author{
Annika Putt <annika@instream.net>
}
