% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_excel.R
\name{export_excel_ORFID}
\alias{export_excel_ORFID}
\title{Export the data frames created with the \code{\link{import_ORFID}} as a xlsx file}
\usage{
export_excel_ORFID(x)
}
\arguments{
\item{x}{PIT_data data frame to be exported.}
}
\value{
Returns an .xlsx file.
}
\description{
Function to export PIT_data data frames as a xlsx file to user working directory
}
\details{
File is saved in the user-specified working directory.
}
\examples{
#export a PIT_data data frame to the working directory.

\dontrun{} export_excel_ORFID(PIT_data)

}
\author{
Hugo Marques
}
