% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_ORFID.R
\name{export_ORFID}
\alias{export_ORFID}
\title{Export the data frames created with the \code{\link{import_ORFID}} as a xlsx file}
\usage{
export_ORFID(x, name, extension)
}
\arguments{
\item{x}{PIT_data data frame to be exported.}
}
\value{
Returns an .xlsx file.
}
\description{
Function to export PIT_data data frames as a xlsx file to user working directory
}
\details{
The file will be saved in the user-specified working directory.
}
\examples{
#  Exporting a PIT_data data frame to the working directory.

#  Import the ORFID reader data to the user working directory
\dontrun{} data_BRA <- import_ORFID("~/Desktop/SCD_BRA.txt")

#  Create the list containing the imported files:
\dontrun{} readers list(data_USA, data_BRA)

#  Combine the files:
\dontrun{} data <- join_multireader_data(readers)

#  Export the multi reader data as a .xlsx file:
\dontrun{} export_excel_ORFID(PIT_data)

}
\seealso{
\code{\link{import_ORFID}} for importing data files from the new generation of Oregon RFID readers
\code{\link{join_multireader_data}} for Combining data from ORFID readers into an multi readers array
}
\author{
Hugo Marques <biohmarques@gmail.com>
}
