% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_ORFID.R
\name{export_ORFID}
\alias{export_ORFID}
\title{Export the data frames created with the \code{\link{import_ORFID}} as a xlsx file}
\usage{
export_ORFID(x, name, extension)
}
\arguments{
\item{x}{PIT_data data frame to be exported.}

\item{name}{a name for the exported file.}

\item{extension}{extension to determine the file type. Should be ".txt", ".csv" or ".xlsx".}
}
\value{
Returns a ".txt", ".csv" or ".xlsx" file depending on the user option.
}
\description{
Function to export PIT_data data frames as a xlsx file to user working directory.
}
\details{
The file will be saved in the user-specified working directory.
}
\examples{
\dontrun{

##  Exporting a PIT_data data frame to the working directory.

##  Create the list containing the imported files:  
readers <- list(reader_1, reader_2)

##  Combine the files:
array <- join_multireader_data(readers)

##  Export the data a .xlsx file:
export_excel_ORFID(x = array, name = "PIT_data", extension = ".xlsx")

##  Export the data as a .txt file:
export_excel_ORFID(x = array, name = "PIT_data", extension = ".txt")

##  Summarizing tag info
tags <- tag_summary(array) 

##  Export the tag info as a .xlsx file:
export_excel_ORFID(x = tags, name = "tags_info", extension = ".xlsx")
}
}
\seealso{
\code{\link{import_ORFID}} for importing data files from the new generation of Oregon RFID readers
\code{\link{join_multireader_data}} for Combining data from ORFID readers into an multi readers array
}
\author{
Hugo Marques <biohmarques@gmail.com>
}
