% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_ORFID.R
\name{export_ORFID}
\alias{export_ORFID}
\title{Export data compiled from Oregon RFID ORMR and ORSR antenna readers}
\usage{
export_ORFID(x, name, extension)
}
\arguments{
\item{x}{data frame to be exported.}

\item{name}{file name.}

\item{extension}{file extension (".csv" or ".xlsx").}
}
\value{
A file is saved in the working directory.
}
\description{
Data compiled using \code{\link{import_ORFID}} or \code{\link{join_multireader_data}} are exported to the working directory as either a .csv or .xlsx file.
}
\details{
The file will be saved in the current working directory. Any data frame created using functions in the ORFID package can be exported using this function.
}
\examples{
\dontrun{

# Exporting a PIT_data data frame to the working directory.

# Create a list containing compiled reader data:
readers <- list(reader_1, reader_2)

# Join data into a multi-reader array:
array <- join_multireader_data(readers)

# Export the data a .xlsx file:
export_ORFID(x = array, name = "PIT_data", extension = ".xlsx")

# Summarize tag info:
tags <- tag_summary(array) 

# Export the tag info as a .xlsx file:
export_ORFID(x = tags, name = "tags_info", extension = ".xlsx")
}
}
\seealso{
\code{\link{import_ORFID}} for importing data files from Oregon RFID ORMR and ORSR antenna readers.

\code{\link{join_multireader_data}} for combining data from Oregon RFID ORMR and ORSR antenna readers into a multi-reader array.
}
\author{
Hugo Marques <biohmarques@gmail.com>
}
