% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/field_names.R
\name{field_names}
\alias{field_names}
\title{Returns the field names from ORSR and ORMR readers}
\usage{
field_names()
}
\value{
A data frame with two variables is returned (Name and Details)
}
\description{
Function to return the field names from ORSR and ORMR readers
}
\details{
Returns the field names and its details from ORSR and ORMR readers data
}
\examples{
\dontrun{

##  Creating a data frame containing the field names from ORSR and ORMR readers.
(FN <- field_names())

}
}
\author{
Hugo Marques <biohmarques@gmail.com>
}
