% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_old_readers.R
\name{import_old_readers}
\alias{import_old_readers}
\title{Import data files from previous generations of Oregon RFID antenna readers (prior to ORMR and ORSR)}
\usage{
import_old_readers(file, delim, verbose = TRUE)
}
\arguments{
\item{file}{.txt file generated by an Oregon RFID reader.}

\item{delim}{field/column delimiter, which must be tab ('\\t'), comma (',') or semi-colon (';').}

\item{verbose}{If TRUE, a description of the compiled data is printed to the console.}
}
\value{
Returns a tibble object.
}
\description{
Function to import files from previous generations of Oregon RFID antenna readers (prior to ORMR and ORSR). Only detections are retained during compilation (events are removed).
}
\details{
The field/column delimiter must be tab, comma or semi-colon for data compilation and further analysis. The function cannot be used for space delimited data. 

The tag number column (TAG) is required for subsequent analyses, and the function will return a warning if TAG is not included in the data file.
}
\examples{
\dontrun{

# Import a single data file from an ORFID reader
old_reader <- import_old_readers("~/Desktop/Old_eader_Data.txt", delim = "\t")

}
}
\author{
Hugo Marques <biohmarques@gmail.com>
}
