% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_old_readers.R
\name{import_old_readers}
\alias{import_old_readers}
\title{Import data files from the old generation of Oregon RFID readers}
\usage{
import_old_readers(file, delim, site_code)
}
\arguments{
\item{file}{txt file generated by an Oregor RFID reader.}

\item{delim}{the field separator character. Values on each line of the file are separated by this character and it must be either tab, comma or semi-colon.}

\item{site_code}{Three letters code for the identify the reader.}
}
\value{
Returns a tibble object.
}
\description{
Function to import files recorded by the first generations of Oregon RFID readers.
}
\details{
A data frame is created in the user environment from a import of a Oregor RFID reader file.
  
the field separator character must be either tab, comma or semi-colon otherwise the function cannot organize properly the variables, generating error messages, without import the file.

The Tag number (TAG) is required for further estatistical analysis. Consider in upload the data from the ORFID reader with variable selected in the Reader field names.
}
\examples{
\dontrun{
Teste
}
}
\author{
Hugo Marques <biohmarques@gmail.com>
}
