% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_ORFID.R
\name{import_ORFID}
\alias{import_ORFID}
\title{Import data files from the new generation of Oregon RFID readers}
\usage{
import_ORFID(file, delim)
}
\arguments{
\item{file}{txt file generated by an Oregor RFID reader.}

\item{delim}{the field separator character. Values on each line of the file are separated by this character.}
}
\description{
Function to import files recorded by the new generation of Oregon RFID readers.
}
\details{
Data frame is created in the user environment.
}
\examples{
#save data frame to a working directory
\dontrun{} import_ORFID("~/Desktop/SCD_BRA.txt")
}
