% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_ORFID.R
\name{import_ORFID}
\alias{import_ORFID}
\title{Import data files from Oregon RFID ORMR and ORSR antenna readers}
\usage{
import_ORFID(file, delim, verbose = TRUE)
}
\arguments{
\item{file}{.txt file generated by an Oregon RFID reader.}

\item{delim}{field/column delimiter, which must be tab ('\\t'), comma (',') or semi-colon (';').}

\item{verbose}{If TRUE, a description of the compiled data is printed to the console.}
}
\value{
Returns a tibble object of data compiled from a single ORFID data file.
}
\description{
Function to import files from Oregon RFID ORMR and ORSR antenna readers. Only detections are retained during compilation (events are removed).
}
\details{
The field/column delimiter must be tab, comma or semi-colon for data compilation and further analysis. The function cannot be used for space delimited data. 

The tag number column (TAG) is required for subsequent analyses, and the function will return a warning if TAG is not included in the data file.
}
\examples{
\dontrun{

# Import a single data file from an ORFID reader
data_BRA <- import_ORFID("~/Desktop/SCD_BRA.txt", delim = "\\\\t")

}
}
\author{
Hugo Marques <biohmarques@gmail.com>, Annika Putt <annika@instream.net>
}
