% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_ORFID_events.R
\name{import_ORFID_events}
\alias{import_ORFID_events}
\title{Import event records from Oregon RFID ORMR and ORSR antenna readers}
\usage{
import_ORFID_events(file, delim, verbose = TRUE)
}
\arguments{
\item{file}{.txt file generated by an Oregon RFID reader.}

\item{delim}{field/column delimiter, which must be tab ('\\t'), comma (',') or semi-colon (';').}

\item{verbose}{If TRUE, a description of the compiled data is printed to the console.}
}
\value{
Returns a tibble object.
}
\description{
Function to import files from Oregon RFID ORMR and ORSR antenna readers. Only event records are retained during compilation (detections are removed).
}
\details{
The field/column delimiter must be tab, comma or semi-colon for data compilation. The function cannot be used for space delimited data.
}
\examples{
\dontrun{

# Importing event records from a single ORFID reader 
data_BRA <- import_ORFID_events("~/Desktop/SCD_BRA.txt")

}
}
\author{
Hugo Marques <biohmarques@gmail.com>
}
