% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_ORFID_events.R
\name{import_ORFID_events}
\alias{import_ORFID_events}
\title{Import event records from the new generation of Oregon RFID readers}
\usage{
import_ORFID_events(file, delim)
}
\arguments{
\item{file}{.txt file generated by an Oregon RFID reader.}

\item{delim}{field separator character, which must be tab (\t), comma, or semi-colon.}
}
\value{
Returns a tibble object.
}
\description{
Function to import event records from files generated by the new generation of Oregon RFID readers.
}
\details{
A data frame is created in the user environment containing only event records (no detection data) from an Oregon RFID reader file.
  
The field separator character must be either tab, comma or semi-colon; space separators are not permitted.
}
\examples{
\dontrun{

##  Importing the event records from the ORFID reader data to the user working directory
data_BRA <- import_ORFID_events("~/Desktop/SCD_BRA.txt")

}
}
\author{
Hugo Marques <biohmarques@gmail.com>
}
