% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marker_tag_plot.R
\name{marker_tag_plot}
\alias{marker_tag_plot}
\title{Ploting marker tag detections of \code{\link{import_ORFID}}}
\usage{
marker_tag_plot(x, tag, gap)
}
\arguments{
\item{x}{data frame to be analyzed.}

\item{tag}{marker tag number. Must be between "".}

\item{gap}{Value in seconds, if you want to filter only values greater than or equal to "gap".}
}
\value{
Returns a plot from the marker tag data.
}
\description{
Function for getting info about the tags recorded in a PIT_data data frame from \code{\link{import_ORFID}} function (number of unique tags, tags type and records per tag).
}
\details{
A plot is created with the data from the marker tag.
}
\examples{
\dontrun{

##  Create the list containing the imported files:
readers <- list(reader_1, reader_2, reader_3)

##  Combine the files:
PIT_data <- join_multireader_data(readers)

##  Ploting marker tag data
mrker_tag_plot(PIT_data, "0000_000000004978")
mrker_tag_plot(PIT_data, "0000_000000004978", 360)
}
}
\seealso{
\code{\link{import_ORFID}} for importing data files from the new generation of Oregon RFID readers
\code{\link{join_multireader_data}} for Combining data from ORFID readers into an multi readers array
}
\author{
Hugo Marques <biohmarques@gmail.com>
}
