% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/site_summary.R
\name{site_summary}
\alias{site_summary}
\title{Summarizing site-specific info of \code{\link{import_ORFID}}}
\usage{
site_summary(x)
}
\arguments{
\item{x}{data frame to be analyzed.}
}
\value{
Returns a tibble object.
}
\description{
Function for getting info about the tsites in a PIT_data data frame from \code{\link{import_ORFID}} function (number of records, number of tags detected and last tag seen per site).
}
\details{
A data frame is created in the user environment with the data grouped by TAG IDs.
}
\examples{
\dontrun{

##  Create the list containing the imported files:
readers <- list(reader_1, reader_2)

##  Combine the files:
PIT_data <- join_multireader_data(readers)

##  Summarizing site info
site_summary(PIT_data)
}
}
\seealso{
\code{\link{import_ORFID}} for importing data files from the new generation of Oregon RFID readers
\code{\link{join_multireader_data}} for Combining data from ORFID readers into an multi readers array
}
\author{
Hugo Marques <biohmarques@gmail.com>
}
